requirements.txt
# ── Python dependencies ───────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
groq==0.9.0
langgraph==0.1.19
langchain-core==0.2.22
requests==2.32.3
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.2

# ── Install ───────────────────────────────────────────────────────────────────
# pip install -r requirements.txt

# ── Run the backend ───────────────────────────────────────────────────────────
# uvicorn main:app --reload --port 8000

# ── React setup (run from /frontend) ─────────────────────────────────────────
# npm create vite@latest frontend -- --template react
# cd frontend
# npm install react-leaflet leaflet axios
# npx tailwindcss init   (then add content: ["./src/**/*.{js,jsx}"] to tailwind.config.js)
# npm run dev            → runs on http://localhost:5173

# ── Environment variables ─────────────────────────────────────────────────────
# Create a .env file in /backend:
#   GROQ_API_KEY=gsk_xxxxxxxxxxxxxxxxxxxx
#   GOOGLE_MAPS_KEY=AIzaxxxxxxxxxxxxxxxxx   (optional for now)
